Editor: Rename codeeditor_widgets.py to gotoline.py
That's now the only widget contained in that module.
