+++ spyder/plugins/editor/widgets/tests/test_shortcuts.py
@@ -20,7 +20,7 @@ from qtpy.QtWidgets import QApplication
 
 # Local imports
 from spyder.config.base import running_in_ci
-from spyder.plugins.editor.widgets.codeeditor_widgets import GoToLineDialog
+from spyder.plugins.editor.widgets.gotoline import GoToLineDialog
 from spyder.plugins.editor.widgets.editor import EditorStack
 from spyder.config.manager import CONF
